+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new OSV option into FlawSource
 - Allow searching by CVE similarity (OSIDB-2482)
 - Add CC lists to Jira trackers and to Bugzilla trackers (OSIDB-2191)
+- Enable flaw draft creation in BZ (OSIDB-2261)
 
 ### Changed
 - Make workflows API RESTful (OSIDB-1716)